SmartInventoryAI-backend-main/requirements.txt
flask
flask-cors
ultralytics
pillow
gunicorn
opencv-python-headless